+++ lib/map.nix
@@ -1,6 +1,6 @@
 { system, lib, inputs, ... }:
 let
-  inherit (builtins) attrValues readDir toString hashString pathExists;
+  inherit (builtins) attrValues mapAttrs readDir typeOf toString hashString pathExists;
   inherit (lib) flatten mapAttrs' mapAttrsToList filterAttrs hasPrefix hasSuffix nameValuePair removeSuffix;
 in rec
 {
@@ -17,7 +17,7 @@ in rec
         in
         if type == "directory" && pathExists "${path}/default.nix"
           then nameValuePair name (func path)
-        else if type == "regular" && name != "default.nix" && name != "repl.nix" && hasSuffix ".nix" name
+        else if type == "regular" && name != "default.nix" && hasSuffix ".nix" name
           then nameValuePair (removeSuffix ".nix" name) (func path)
         else nameValuePair "" null)
       (readDir dir);
@@ -27,7 +27,12 @@ in rec
   {
     name = "patched-input-${hashString "md5" (toString pkgs)}";
     src = pkgs;
-    inherit patches;
+    patches = if typeOf patches == "list" then patches
+    else flatten (mapAttrsToList (name: type:
+      if hasSuffix ".diff" name || hasSuffix ".patch" name
+        then patches + "/${name}"
+      else null)
+    (readDir patches));
   })
   {
     inherit system;
@@ -39,19 +44,24 @@ in rec
     };
   };
 
-  # Patch Mapping Function
-  patches = dir: flatten (mapAttrsToList (name: type:
-  if hasSuffix ".diff" name || hasSuffix ".patch" name
-    then dir + "/${name}"
-  else null)
-  (readDir dir));
+  # Nix Input Mapping Functions
+  nix =
+  {
+    registry = mapAttrs (name: value: { flake = value; }) (filterAttrs (name: value: value ? outputs) inputs);
+    etc = mapAttrs' (name: value: { name = "nix/inputs/${name}"; value = { source = value.outPath; }; }) inputs;
+    path = mapAttrsToList (name: value: "${name}=/etc/nix/inputs/${name}") (filterAttrs (name: value: value.outputs ? legacyPackages || value.outputs ? packages) (filterAttrs (name: value: value ? outputs) inputs));
+  };
 
   # Secrets Mapping Function
   secrets = dir: choice:
-    mapAttrs' (name: type: (nameValuePair name
-    {
-      sopsFile =  dir + "/${name}";
-      format = "binary";
-      neededForUsers = choice;
-    })) (readDir dir);
+    filter
+      (name: type: type != null && !(hasPrefix "_" name))
+      (name: type:
+        (nameValuePair name
+        {
+          sopsFile =  dir + "/${name}";
+          format = "binary";
+          neededForUsers = choice;
+        }))
+      (readDir dir);
 }