+++ modules/nix/default.nix
@@ -1,13 +1,21 @@
-{ inputs, pkgs, ... }:
-rec
+{ util, pkgs, ... }:
+let
+  inherit (util) map;
+in rec
 {
   imports = [ ./cachix.nix ];
 
   ## Nix Settings ##
   config =
   {
+    # Nix Path
+    environment.etc = map.nix.etc;
+    nix.nixPath = map.nix.path;
+    nix.registry = map.nix.registry;
+
     nix =
     {
+      # Version
       package = pkgs.nix_2_4;
 
       # Garbage Collection
@@ -19,24 +27,12 @@ rec
         options   = "--delete-older-than 7d";
       };
 
-      # Nix Path
-      nixPath = [ "nixpkgs=${inputs.nixpkgs}" ];
-
       # User Permissions
       trustedUsers = [ "root" "@wheel" ];
 
       # Additional Features
       extraOptions = "experimental-features = nix-command flakes recursive-nix";
       systemFeatures = [ "kvm" "recursive-nix" ];
-
-      # Flakes
-      registry =
-      {
-        self.flake = inputs.self;
-        nixpkgs.flake = inputs.nixpkgs;
-        unstable.flake = inputs.unstable;
-        home-manager.flake = inputs.home;
-      };
     };
   };
 }